requirements.txt
asgiref==3.11.0
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gitdb==4.0.12
GitPython==3.1.41
PyJWT==2.10.1
python-dotenv==1.2.1
setuptools==75.6.0
smmap==5.0.2
sqlparse==0.5.5
